ress#6 Details Button for Products for Sale
When a user is logged in, he can access the page of details of a product for sale, decide how many items of this product he is going to buy and add them to the shopping cart. There is a bug: if the user is not logged in and press the details button, it should redirect him to the log in page; but instead, it is redirecting him to a blank page.
Also the Add to Cart button hasn't been implemented yet, it doesn't do anything yet