refs #24 - Message in the Summary View
A string Message was added to the Order Details Cart View Model. If this string is empty it will not be displayed in the Summary View. In case, it is not empty this message will inform the user some changes were applied to his shopping cart because of availability of some products